Split the builder into smaller files
