+++ Lib/glyphsLib/builder/blue_values.py
@@ -1,0 +1,36 @@
+# Copyright 2015 Google Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import (print_function, division, absolute_import,
+                        unicode_literals)
+
+
+def to_ufo_blue_values(_context, ufo, master):
+    """Set postscript blue values from Glyphs alignment zones."""
+
+    alignment_zones = master.alignmentZones
+    blue_values = []
+    other_blues = []
+    for zone in sorted(alignment_zones):
+        pos = zone.position
+        size = zone.size
+        val_list = blue_values if pos == 0 or size >= 0 else other_blues
+        val_list.extend(sorted((pos, pos + size)))
+
+    ufo.info.postscriptBlueValues = blue_values
+    ufo.info.postscriptOtherBlues = other_blues
+
+
+def to_glyphs_blue_values(context, ufo, master):
+    pass
