+++ Lib/glyphsLib/builder/common.py
@@ -1,0 +1,23 @@
+# Copyright 2015 Google Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import (print_function, division, absolute_import,
+                        unicode_literals)
+
+UFO_FORMAT = '%Y/%m/%d %H:%M:%S'
+
+
+def to_ufo_time(datetime_obj):
+    """Format a datetime object as specified for UFOs."""
+    return datetime_obj.strftime(UFO_FORMAT)
