+++ Lib/glyphsLib/interpolation.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 from __future__ import (print_function, division, absolute_import,
                         unicode_literals)
 
@@ -21,7 +20,8 @@ import logging
 import os
 import xml.etree.ElementTree as etree
 
-from glyphsLib.builder.ufo import set_custom_params, build_stylemap_names
+from glyphsLib.builder.custom_params import set_custom_params
+from glyphsLib.builder.names import build_stylemap_names
 from glyphsLib.builder.constants import GLYPHS_PREFIX
 
 from glyphsLib.util import build_ufo_path, write_ufo, clean_ufo, clear_data
