+++ tests/roundtrip_test.py
@@ -15,28 +15,27 @@
 # limitations under the License.
 
 import unittest
-import difflib
-import sys
+import os
 
+import glyphsLib
 from glyphsLib import classes
-from glyphsLib.builder import (to_ufos, to_glyphs)
 
 import test_helpers
 
 
-class ClassRoundtripTest(unittest.TestCase, test_helpers.AssertLinesEqual):
-    def assertRoundtrip(self, font):
-        expected = test_helpers.write_to_lines(font)
-        roundtrip = to_glyphs(to_ufos(font))
-        actual = test_helpers.write_to_lines(roundtrip)
-        self.assertLinesEqual(
-            expected, actual,
-            "The font has been modified by the roundtrip")
-
+class UFORoundtripTest(unittest.TestCase, test_helpers.AssertUFORoundtrip):
     def test_empty_font(self):
         empty_font = classes.GSFont()
         empty_font.masters.append(classes.GSFontMaster())
-        self.assertRoundtrip(empty_font)
+        self.assertUFORoundtrip(empty_font)
+
+    def test_GlyphsUnitTestSans(self):
+        self.skipTest("TODO")
+        filename = os.path.join(os.path.dirname(__file__),
+                                'data/GlyphsUnitTestSans.glyphs')
+        with open(filename) as f:
+            font = glyphsLib.load(f)
+        self.assertUFORoundtrip(font)
 
 
 if __name__ == '__main__':
